+++ lib/dbus/bus.rb
@@ -76,6 +76,7 @@ module DBus
     def export(obj)
       obj.service = self
       get_node(obj.path, create: true).object = obj
+      object_manager_for(obj)&.object_added(obj)
     end
 
     # Undo exporting an object _obj_.
@@ -93,6 +94,7 @@ module DBus
       parent_node = get_node(parent_path, create: false)
       return false unless parent_node
 
+      object_manager_for(obj)&.object_removed(obj)
       obj.service = nil
       parent_node.delete(node_name).object
     end
@@ -112,10 +114,30 @@ module DBus
         end
         n = n[elem]
       end
-      if n.nil?
-        DBus.logger.debug "Warning, unknown object #{path}"
-      end
       n
+    end
+
+    # Find the (closest) parent of *object*
+    # implementing the ObjectManager interface, or nil
+    # @return [DBus::Object,nil]
+    def object_manager_for(object)
+      path = object.path
+      node_chain = get_node_chain(path)
+      om_node = node_chain.reverse_each.find do |node|
+        node.object&.is_a? DBus::ObjectManager
+      end
+      om_node&.object
+    end
+
+    # All objects (not paths) under this path (except itself).
+    # @param path [ObjectPath]
+    # @return [Array<DBus::Object>]
+    # @raise ArgumentError if the *path* does not exist
+    def descendants_for(path)
+      node = get_node(path, create: false)
+      raise ArgumentError, "Object path #{path} doesn't exist" if node.nil?
+
+      node.descendant_objects
     end
 
     #########
@@ -123,6 +145,19 @@ module DBus
     private
 
     #########
+
+    # @raise ArgumentError if the *path* does not exist
+    def get_node_chain(path)
+      n = @root
+      result = [n]
+      path.sub(%r{^/}, "").split("/").each do |elem|
+        n = n[elem]
+        raise ArgumentError, "Object path #{path} doesn't exist" if n.nil?
+
+        result.push(n)
+      end
+      result
+    end
 
     # Perform a recursive retrospection on the given current _node_
     # on the given _path_.
@@ -198,6 +233,15 @@ module DBus
                              .join(",")
       "#{s}{#{contents_sub_inspect}}"
     end
+
+    # All objects (not paths) under this path (except itself).
+    # @return [Array<DBus::Object>]
+    def descendant_objects
+      children_objects = values.map(&:object).compact
+      descendants = values.map(&:descendant_objects)
+      flat_descendants = descendants.reduce([], &:+)
+      children_objects + flat_descendants
+    end
   end
 
   # FIXME: rename Connection to Bus?
@@ -233,7 +277,6 @@ module DBus
       @method_call_msgs = {}
       @signal_matchrules = {}
       @proxy = nil
-      @object_root = Node.new("/")
     end
 
     # Dispatch all messages that are available in the queue,